Version 10
This version covers the test_1.c through test_3.c cases and test_5.c, and is partially functional for test_4.c
One error that must still be fixed is in the case of test_4.c, if strlen() < 16 is in the if statement and strcpy() is in the else statement (meaning it is not guarded) the addon marks it as guarded. I am also planning to add line number checking to make sure strlen() and the "<" operator are on the same line not just in the same overall scope.